hude

+++ lib/view/setting/setting_page.dart
@@ -7,6 +7,8 @@ import 'package:fiber/view/setting/components/custom_support_card.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_animate/flutter_animate.dart';
 
+import '../../controller/set_data_controller.dart';
+
 class SettingPage extends StatefulWidget {
   const SettingPage({super.key});
 
@@ -27,9 +29,11 @@ class _SettingPageState extends State<SettingPage> {
           SizedBox(
             height: Insets.exLarge + 24,
           ),
-          CustomProfileCard(
-            userModel: homeController.userInfo.value,
-          ),
+          homeController.userInfo.value.id == null
+              ? Container()
+              : CustomProfileCard(
+                  userModel: homeController.userInfo.value,
+                ),
           SizedBox(
             height: Insets.medium,
           ),
@@ -41,45 +45,53 @@ class _SettingPageState extends State<SettingPage> {
             height: Insets.small,
           ),
           const CustomSupportCard(),
-          SizedBox(
-            height: Insets.medium,
-          ),
-          Text(
-            'الملف الشخصي'.tr,
-            style: Theme.of(context).textTheme.titleMedium,
-          ),
-          SizedBox(
-            height: Insets.small,
-          ),
-          ListTile(
-            contentPadding: EdgeInsets.zero,
-            dense: true,
-            visualDensity: VisualDensity.comfortable,
-            enableFeedback: false,
-            minVerticalPadding: 0,
-            titleAlignment: ListTileTitleAlignment.center,
-            style: ListTileStyle.list,
-            horizontalTitleGap: Insets.small,
-            title: Text(
-              'تغيير كلمة السر'.tr,
-              style: Theme.of(context).textTheme.titleMedium!.copyWith(
-                    color: Theme.of(context).colorScheme.outline,
-                  ),
-            ),
-            leading: SvgPicture.asset(
-              Assets.assetsIconsPassword,
-              width: 24,
-              height: 24,
-            ),
-            trailing: SvgPicture.asset(
-              Assets.assetsIconsCaretRight,
-              width: 24,
-              height: 24,
-            ),
-          ),
-          SizedBox(
-            height: Insets.medium,
-          ),
+          // homeController.userInfo.value.id == null
+          //     ? Container()
+          //     : SizedBox(
+          //         height: Insets.medium,
+          //       ),
+          // homeController.userInfo.value.id == null
+          //     ? Container()
+          //     : Text(
+          //         'الملف الشخصي'.tr,
+          //         style: Theme.of(context).textTheme.titleMedium,
+          //       ),
+          // homeController.userInfo.value.id == null
+          //     ? Container()
+          //     : SizedBox(
+          //         height: Insets.small,
+          //       ),
+          // homeController.userInfo.value.id == null
+          //     ? Container()
+          //     : ListTile(
+          //         contentPadding: EdgeInsets.zero,
+          //         dense: true,
+          //         visualDensity: VisualDensity.comfortable,
+          //         enableFeedback: false,
+          //         minVerticalPadding: 0,
+          //         titleAlignment: ListTileTitleAlignment.center,
+          //         style: ListTileStyle.list,
+          //         horizontalTitleGap: Insets.small,
+          //         title: Text(
+          //           'تغيير كلمة السر'.tr,
+          //           style: Theme.of(context).textTheme.titleMedium!.copyWith(
+          //                 color: Theme.of(context).colorScheme.outline,
+          //               ),
+          //         ),
+          //         leading: SvgPicture.asset(
+          //           Assets.assetsIconsPassword,
+          //           width: 24,
+          //           height: 24,
+          //         ),
+          //         trailing: SvgPicture.asset(
+          //           Assets.assetsIconsCaretRight,
+          //           width: 24,
+          //           height: 24,
+          //         ),
+          //       ),
+          // SizedBox(
+          //   height: Insets.medium,
+          // ),
           Text(
             'الحساب'.tr,
             style: Theme.of(context).textTheme.titleMedium,
@@ -94,16 +106,26 @@ class _SettingPageState extends State<SettingPage> {
             enableFeedback: false,
             minVerticalPadding: 0,
             onTap: () {
-              prefs.clear();
-              Get.offAll(() => const LoginPage());
+              if (homeController.userInfo.value.id == null) {
+                Get.to(() => const LoginPage());
+                Get.find<SetDataController>().page.value = 0;
+              } else {
+                Get.find<SetDataController>().page.value = 0;
+                prefs.clear();
+                Get.offAll(() => const LoginPage());
+              }
             },
             titleAlignment: ListTileTitleAlignment.center,
             style: ListTileStyle.list,
             horizontalTitleGap: Insets.small,
             title: Text(
-              'تسجيل الخروج'.tr,
+              homeController.userInfo.value.id == null
+                  ? "تسجيل الدخول"
+                  : 'تسجيل الخروج'.tr,
               style: Theme.of(context).textTheme.titleMedium!.copyWith(
-                    color: Theme.of(context).colorScheme.error,
+                    color: homeController.userInfo.value.id == null
+                        ? Theme.of(context).colorScheme.primary
+                        : Theme.of(context).colorScheme.error,
                   ),
             ),
             leading: SvgPicture.asset(
@@ -115,7 +137,9 @@ class _SettingPageState extends State<SettingPage> {
               Assets.assetsIconsCaretRight,
               width: 24,
               height: 24,
-              color: context.theme.colorScheme.error,
+              color: homeController.userInfo.value.id == null
+                  ? Theme.of(context).colorScheme.primary
+                  : context.theme.colorScheme.error,
             ),
           ),
         ].animate(interval: 50.ms).fade(),
